requirements.txt
flask
flask-socketio
flask-pymongo
marshmallow
pymongo
python-dotenv
gevent
gevent-websocket
pymongo-migrate
flask-cors

